[File&Code]: TaskDetail + EllipsisMenu & Deletion
- TaskDetail 삭제기능 구현
- EllipsisMenu 파일 추가
+ 함수형 파라미터 생성
- TaskInfo&TaskTimer&CompletionMemo 코드 수정
task를 파라미터로 입력받아 해당 task의 정보를 표시
