+++ Boong-To-Do/Model/TaskModel.swift
@@ -17,9 +17,10 @@ struct UserInfo {
 struct Todo: Identifiable {
     let id = UUID()
     var title: String
-    var content: String?
+    var content: String = ""
     var requiredTime: Int
     var executedTime: Int = 0
+    // TODO: 날짜 지정입력 하기
     var createdAt: Date
     var finishedAt: Date?
     var memo: [Memo]?
