+++ Boong-To-Do/View/TaskDetailView.swift
@@ -12,12 +12,21 @@ import SwiftUI
  할일 상세목록 + 타이머
  */
 struct TaskDetailView: View {
+    
+    @EnvironmentObject var viewModel: TaskViewModel
+    @State var task: Todo
+    
     var body: some View {
         VStack {
         // TODO: 뷰에서 표시할 데이터를 전달해야함
-            TaskInfo()
+            HStack {
+                Spacer()
+                EllipsisMenu(action: { viewModel.deleteTask(id: task.id) })
+            }
             
-            TaskTimer()
+            TaskInfo(task: $task)
+            
+            TaskTimer(task: $task)
             
         }
         .padding(.horizontal, 20)
@@ -25,5 +34,5 @@ struct TaskDetailView: View {
 }
 
 #Preview {
-    TaskDetailView()
+    TaskDetailView(task: Todo(title: "디테일뷰", requiredTime: 12, createdAt: Date.now))
 }
